smart_contracts/requirements.txt
beaker-py==2.5.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
docstring-parser==0.14.1
executing==2.0.1
google-crc32c==1.8.0
grpcio==1.78.0
idna==3.11
msgpack==1.1.2
packaging==26.0
protobuf==6.33.5
py-algorand-sdk==2.11.1
pycparser==3.0
pycryptodomex==3.23.0
PyNaCl==1.6.2
pyteal==0.27.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
semantic-version==2.10.0
tabulate==0.9.0
typing_extensions==4.15.0
urllib3==2.6.3
